requirements.txt
torch>=1.7.0
opencv-python>=4.5.3
numpy>=1.19.5
Pillow>=8.0.0